# docs/swagger.yaml
basePath: /
definitions:
  handlers.ClusterResponse:
    properties:
      config:
        $ref: '#/definitions/models.ClusterConfig'
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: clstr_abc123
        type: string
      max_servers:
        example: 10
        type: integer
      min_servers:
        example: 2
        type: integer
      name:
        example: production-cluster
        type: string
      status:
        example: active
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  models.ClusterConfig:
    properties:
      collector_endpoint:
        type: string
      target_cpu:
        type: number
    type: object
host: localhost:8080
info:
  contact:
    email: tanmaytambat01@gmail.com
    name: API Support
  description: A cloud infrastructure autoscaling service that monitors cluster metrics
    and automatically scales resources based on demand.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cloud Autoscaler API
  version: "1.0"
paths:
...
  /clusters:
    get:
      description: Get all clusters owned by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of clusters
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List clusters
      tags:
      - Clusters
  /clusters/{id}:
    get:
      description: Get a specific cluster by ID
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cluster details
          schema:
            $ref: '#/definitions/handlers.ClusterResponse'
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cluster not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cluster
      tags:
      - Clusters
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"